Move Surface to utils
Surface was not imported in utils, which was breaking coord_to_label.
I could not perform a relative import of io from utils (I do not understand
that bug, but there were some other reports of similar issues with package
modules named io on Google).

Anyway, tests pass and docs build with the reorganization.