Remove config_opts and references to config file in surfer.viz
The Brain object can now be directly configured with keyword argument, although
`config_opts` is still caught and used for backwards compatability.

This commit does not fully remove the config file code, and it has not updated
the example scripts to remove config_opts, but the unit tests should pass.