Fix errors that were crashing doc build
GH-#115 removed surfer.io.read_label but did not change the example
that was referencing it. I'm surprised that got through Travis--
does Travis not build the docs?

While fixing this I also noticed a couple of things that had gotten
out of date and fixed an oversight from the config file removal.