allowing transparancy in the centre of diverging color maps
There was a mechanism for smoothly fading out small values of some data shown on
the brain, but this did not work for diverging colormaps for which 'small'
values are in the centre of the color map instead of on the lower edge. I have
added a new flag "divergent" to scale_data_colormap which tells the function to
assume a divergent colormap and apply transparency accordingly. Compared to
adding separate overlays for the two sides of the diverging colormap this allows
easier, interactive exploration of suitable colormap settings.

I have adapted the example plot_resting_correlations to demonstrate the effect.