allow fully transparent center of divergent colormaps + new default cmaps
Added this feature according to discussion on
https://github.com/nipy/PySurfer/pull/207

fmin now gives the distance from center at which colors should first be shown. I
implemented this by inserting a fully transparent part in the centre of the
colormap. Because the scalar_lut_manager of Mayavi and eventually VTK chooses
colours based on linear interpolation, you cannot map a range of values to a
single color at the centre of the colormap. Thus, I add new transparent colors
in the centre in proportion to the size of the corresponding value range. This
also ensures that the colorbar automatically gives meaningful information.

add_data now chooses colormaps automatically, if cmap="auto" is chosen (the
default). The decision is currently only based on whether the center argument
indicates a divergent colormap, see commit b5ff3be0. This meant that I had to
add center as an argument to add_data and decided to complete the interface for
scale_data_colormap by adding also new arguments mid and transparent to
add_data which saves a subsequent call to scale_data_colormap. I'm unsure about
the most compatible order of these arguments, but so far didn't find any parts
of the code which used the order of arguments beyond min and max.