[FIX] strange coloring of extreme values when rescaling divergent colormap
Somehow Mayavi mapped values beyond the lookup-table range to inconsistent
colors. Specifically, when rescaling a divergent colormap from fmin=0.1 to
fmin=0, the values beyond +/-fmax got strange colors that apparently depended on
the previous setting of fmin, or its relation to fmax. I traced this to a
problem with changing colormap sizes which Mayavi apparently doesn't handle
properly, but I couldn't figure out exactly where the problem was, i.e., which
VTK attributes did not get set consistently when changing the colormap size.
Nevertheless, the issue can be avoided by simply sticking to the standard
colormap size of 256. I originally did not want to do that to maintain the color
resolution of 128 colors for the supposedly interesting values between fmin and
fmax, but I guess we can live with a lower resolution, if colors are displayed
consistently.