accounting for divergent colormaps with jumps in the center
Some divergent colormaps sharply transition from one color to another in the
center of the colormap. For these the colorbar looked strange, because the first
color of the second half of the colormap was chosen to fill up the center of the
colormap for fmin>0. We now identify these colormaps and use a neutral grey as
fill color for them instead. For colormaps with smooth transitions nothing
should change.