correct functionality of fmin and transparent colorbar
I previously set all values below fmin or within [center-fmin, center+fmin] to
fully transparent irrespective of the setting of 'transparent'. This is not the
functionality of fmin as previously defined for sequential colormaps. fmin only
defines all values that should be mapped to the left edge of the colormap, or
the centre colours in the colormap. I have now corrected that such that, when
transparent=False, all data values are displayed with the corresponding colours.

To not display small values also with transparent=False add_data has
additionally the 'thresh' argument which uses a Mayavi threshold filter to clip
data that falls below the given threshold. Unfortunately, I couldn't find an
appropriate VTK filter that can clip data in the centre of the data range, yet.
So I left this part of the interface untouched.

I found, however, an option to switch on transparency of the colorbar and now
toggle this together with 'transparent'.